[test] AndroidTest: refactor packages
