+++ app/src/androidTest/java/com/rwmobi/giphytrending/ui/test/CustomTestRunner.kt
@@ -3,7 +3,7 @@
  * https://github.com/ryanw-mobile
  */
 
-package com.rwmobi.giphytrending.ui.utils
+package com.rwmobi.giphytrending.ui.test
 
 import android.app.Application
 import android.content.Context
