+++ app/src/androidTest/java/com/rwmobi/giphytrending/ui/test/GiphyTrendingTestRule.kt
@@ -3,7 +3,7 @@
  * https://github.com/ryanw-mobile
  */
 
-package com.rwmobi.giphytrending.ui
+package com.rwmobi.giphytrending.ui.test
 
 import androidx.compose.ui.test.junit4.AndroidComposeTestRule
 import androidx.test.ext.junit.rules.ActivityScenarioRule
