+++ app/src/androidTest/java/com/rwmobi/giphytrending/ui/test/SemanticsMatchers.kt
@@ -3,7 +3,7 @@
  * https://github.com/ryanw-mobile
  */
 
-package com.rwmobi.giphytrending.ui.utils
+package com.rwmobi.giphytrending.ui.test
 
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.SemanticsProperties
